+++ Samples/BasicArchiveSample/BasicArchiveSample/Commands/CommandType.cs
@@ -1,0 +1,16 @@
+﻿namespace BasicArchiveSample.Commands
+{
+    public enum CommandType
+    {
+        None,
+        Invalid,
+        Export,
+        ExportAll,
+        ExportMatching,
+        Help,
+        Exit,
+        Info,
+        Find,
+        Drop,
+    }
+}