+++ Source/Ba2Tools/BA2Archive.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
+using System.Threading;
 
 namespace Ba2Tools
 {
@@ -16,11 +15,17 @@ namespace Ba2Tools
         /// <summary>
         /// Archive version defined in header.
         /// </summary>
+        /// <value>
+        /// The version.
+        /// </value>
         public UInt32 Version { get { return Header.Version; } }
 
         /// <summary>
         /// Number of files stored in archive.
         /// </summary>
+        /// <value>
+        /// The total files.
+        /// </value>
         public UInt32 TotalFiles { get { return Header.TotalFiles; } }
 
         /// <summary>
@@ -29,20 +34,68 @@ namespace Ba2Tools
         protected internal UInt64 NameTableOffset { get { return Header.NameTableOffset; } }
 
         /// <summary>
-        /// Path to file that was opened.
+        /// Gets the archive stream.
         /// </summary>
-        //public string FilePath { get; internal set; }
-
+        /// <value>
+        /// The archive stream.
+        /// </value>
         public Stream ArchiveStream { get; internal set; }
 
+        /// <summary>
+        /// Gets the archive header.
+        /// </summary>
+        /// <value>
+        /// The archive header.
+        /// </value>
         public BA2Header Header { get; internal set; }
 
         /// <summary>
         /// ListFiles() cache.
         /// </summary>
-        protected string[] _fileListCache = null;
+        /// <seealso cref="ListFiles(bool)"/>
+        protected List<string> _fileListCache = null;
 
+        ~BA2Archive()
+        {
+            Dispose(false);
+        }
+
+        #region Extract methods
+        /// <summary>
+        /// Extract all files from archive to specified directory.
+        /// </summary>
+        /// <param name="destination">Destination directory where extracted files will be placed.</param>
+        /// <param name="overwriteFiles">Overwrite files on disk with extracted ones?</param>
         public virtual void ExtractAll(string destination, bool overwriteFiles = false)
+        {
+            this.ExtractAll(destination, CancellationToken.None, null, overwriteFiles);
+        }
+
+        /// <summary>
+        /// Extract all files from archive to specified directory with
+        /// cancellation token.
+        /// </summary>
+        /// <param name="destination">Directory where extracted files will be placed.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="overwriteFiles">Overwrite existing files in extraction directory?</param>
+        public virtual void ExtractAll(string destination, CancellationToken cancellationToken, bool overwriteFiles = false)
+        {
+            this.ExtractAll(destination, cancellationToken, null, overwriteFiles);
+        }
+
+        /// <summary>
+        /// Extract all files from archive to specified directory with
+        /// cancellation token and progress reporter.
+        /// </summary>
+        /// <param name="destination">Absolute or relative directory path directory where extracted files will be placed.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="progress">Progress reporter ranged from 0 to archive's total files count.</param>
+        /// <param name="overwriteFiles">Overwrite files on disk with extracted ones?</param>
+        public virtual void ExtractAll(
+            string destination,
+            CancellationToken cancellationToken,
+            IProgress<int> progress,
+            bool overwriteFiles = false)
         {
             throw new NotSupportedException("Cannot extract any files because archive type is unknown.");
         }
@@ -50,14 +103,57 @@ namespace Ba2Tools
         /// <summary>
         /// Extract all files from archive.
         /// </summary>
-        /// <seealso cref="BA2ExtractionException"/>
+        /// <param name="fileNames">Files to extract.</param>
         /// <param name="destination">Directory where extracted files will be placed.</param>
         /// <param name="overwriteFiles">Overwrite existing files in extraction directory?</param>
-        public virtual void ExtractFiles(string[] fileNames, string destination, bool overwriteFiles = false)
+        public virtual void ExtractFiles(IEnumerable<string> fileNames, string destination, bool overwriteFiles = false)
+        {
+            this.ExtractFiles(fileNames, destination, CancellationToken.None, null, overwriteFiles);
+        }
+
+        /// <summary>
+        /// Extract all files from archive.
+        /// </summary>
+        /// <param name="fileNames">Files to extract.</param>
+        /// <param name="destination">Directory where extracted files will be placed.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="overwriteFiles">Overwrite existing files in extraction directory?</param>
+        public virtual void ExtractFiles(
+            IEnumerable<string> fileNames,
+            string destination,
+            CancellationToken cancellationToken,
+            bool overwriteFiles = false)
+        {
+            this.ExtractFiles(fileNames, destination, CancellationToken.None, null, overwriteFiles);
+        }
+
+        /// <summary>
+        /// Extract all files from archive to specified directory
+        /// with cancellation token and progress reporter.
+        /// </summary>
+        /// <param name="fileNames">Files to extract.</param>
+        /// <param name="destination">Absolute or relative directory path where extracted files will be placed.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="progress">Progress reporter ranged from 0 to <c>fileNames.Count()</c>.</param>
+        /// <param name="overwriteFiles">Overwrite existing files in extraction directory?</param>
+        public virtual void ExtractFiles(
+            IEnumerable<string> fileNames,
+            string destination,
+            CancellationToken cancellationToken,
+            IProgress<int> progress,
+            bool overwriteFiles = false)
         {
             throw new NotSupportedException("Cannot extract any files because archive type is unknown.");
         }
 
+        /// <summary>
+        /// Extract file contents to stream.
+        /// </summary>
+        /// <param name="fileName">File name or file path from archive.</param>
+        /// <param name="stream">The stream.</param>
+        /// <returns>
+        /// Success is true, failure is false.
+        /// </returns>
         public virtual bool ExtractToStream(string fileName, Stream stream)
         {
             throw new NotSupportedException("Cannot extract any files because archive type is unknown.");
@@ -73,14 +169,16 @@ namespace Ba2Tools
         {
             throw new NotSupportedException("Cannot extract any files because archive type is not known.");
         }
+        #endregion
 
         /// <summary>
-        /// Lists all files in archive.
+        /// Shows all file paths in archive.
         /// </summary>
-        /// <see cref="ExtractFile(string, string)"/>
-        /// <param name="forceListFiles">Force list of files in archive instead of returning cached copy.</param>
-        /// <returns>Array of file paths</returns>
-        public virtual string[] ListFiles(bool forceListFiles = false)
+        /// <param name="forceListFiles">Force list files in archive instead of returning cached copy.</param>
+        /// <returns>
+        /// List of file paths in archive.
+        /// </returns>
+        public virtual IList<string> ListFiles(bool forceListFiles = false)
         {
             if (_fileListCache != null && forceListFiles == false)
                 return _fileListCache;
@@ -91,7 +189,7 @@ namespace Ba2Tools
                 goto invalidNameTableProviden;
             }
 
-            List<string> strings = new List<string>();
+            List<string> strings = new List<string>(Math.Min(10000, (int)TotalFiles));
 
             ArchiveStream.Seek((long)NameTableOffset, SeekOrigin.Begin);
             using (var reader = new BinaryReader(ArchiveStream, Encoding.ASCII, leaveOpen: true))
@@ -114,33 +212,40 @@ namespace Ba2Tools
             // Is all files in archive were listed? (excepted "FileCount" files)
             System.Diagnostics.Debug.Assert(TotalFiles == strings.Count);
 
-            _fileListCache = strings.ToArray();
+            _fileListCache = strings;
             return _fileListCache;
 
-            /// goto case when invalid name table offset was providen
-            invalidNameTableProviden: {
-                _fileListCache = new string[0];
+        /// goto case when invalid name table offset was providen
+        invalidNameTableProviden:
+            {
+                _fileListCache = new List<string>(0);
                 return _fileListCache;
             }
         }
 
+        /// <summary>
+        /// Preloads the data.
+        /// </summary>
+        /// <param name="reader">The reader.</param>
         internal virtual void PreloadData(BinaryReader reader)
         {
             // No data to preload.
         }
 
+        /// <summary>
+        /// Check file existance in archive.
+        /// </summary>
+        /// <param name="fileName">File to check in archive</param>
+        /// <returns></returns>
+        /// <remarks>
+        /// Case-insensitive.
+        /// </remarks>
         public virtual bool ContainsFile(string fileName)
         {
             if (_fileListCache == null)
                 ListFiles();
 
-            for (int i = 0; i < _fileListCache.Length; i++)
-            {
-                if (fileName.Equals(_fileListCache[i], StringComparison.OrdinalIgnoreCase))
-                    return true;
-            }
-
-            return false;
+            return _fileListCache.Contains(fileName, StringComparer.OrdinalIgnoreCase);
         }
 
         /// <summary>
@@ -149,7 +254,20 @@ namespace Ba2Tools
         /// </summary>
         public void Dispose()
         {
-            ArchiveStream.Dispose();
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                if (ArchiveStream != null)
+                {
+                    ArchiveStream.Dispose();
+                    ArchiveStream = null;
+                }
+            }
         }
     }
 }