+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## Version 0.2.0.0
+## Version 0.3
+* More generic method declarations (string[] => IEnumerable<string>)
+* File entries are classes now instead of structs.
+* Added possibility of cancelling extraction and reporting progress of extraction.
+
+## Version 0.2
 * Upgraded to .NET Framework 4.5
 * Added Doxygen support.
 * Added tests.
@@ -10,6 +15,5 @@
 * BA2Archives are distinct from file system now: they can be constructed from plain streams now.
 * BA2Archives now holding a stream to archive data and implements ```IDisposable``` interface. Now you should call ```Dispose()``` when BA2Archive instance is not needed anymore.
 
-## Version 0.1.0.0
-* Initial release.
+## Version 0.1
 * Partial support of general archives: exported packed files are invalid.